Apply MET corrections to all configs, w/phi corrections
The MET corrections (including latest Txy ie phi) are
now applied by default to all MC, for both electron and muon configs.

Since the phi corrections are run-dependent in data, the
process.source has been moved to the top of the file so we can use it to
extract the run period.

This breaks the MET systematics in MC, since it used to use the
shiftedPt() of the main MET object, whereas we now have lots of MET
objects to deal with. This will be fixed in a later commit.